rename OptimizerJax to Optimizer
