+++ tests/test_math/test_compactFock.py
@@ -9,7 +9,7 @@ from mrmustard import math
 from mrmustard.lab import DM, Ggate, SqueezedVacuum, Vacuum
 from mrmustard.lab.transformations.attenuator import Attenuator
 from mrmustard.physics import gaussian
-from mrmustard.training import OptimizerJax
+from mrmustard.training import Optimizer
 
 
 def test_compactFock_diagonal():
@@ -87,7 +87,7 @@ def test_compactFock_diagonal_gradients():
         p = probs[n1]
         return -math.real(p)
 
-    opt = OptimizerJax(symplectic_lr=0.5)
+    opt = Optimizer(symplectic_lr=0.5)
     (G,) = opt.minimize(cost_fn, by_optimizing=[G], max_steps=5)
     for i in range(2, min(20, len(opt.opt_history))):
         assert opt.opt_history[i - 1] >= opt.opt_history[i]
@@ -119,7 +119,7 @@ def test_compactFock_1leftover_gradients():
             *SqueezedVacuum(0, r=1).phase_space(0)[:2],
         )
 
-    opt = OptimizerJax(symplectic_lr=0.1)
+    opt = Optimizer(symplectic_lr=0.1)
     (G,) = opt.minimize(cost_fn, by_optimizing=[G], max_steps=5)
     for i in range(2, len(opt.opt_history)):
         assert opt.opt_history[i - 1] >= opt.opt_history[i]
