+++ tests/test_training/test_optimizer.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Tests for the OptimizerJax class"""
+"""Tests for the Optimizer class"""
 
 import numpy as np
 import pytest
@@ -40,13 +40,13 @@ from mrmustard.lab import (
 )
 from mrmustard.math.parameters import Variable
 from mrmustard.physics.gaussian import number_means, von_neumann_entropy
-from mrmustard.training import OptimizerJax
+from mrmustard.training import Optimizer
 
 
 @pytest.mark.requires_backend("jax")
-class TestOptimizerJax:
+class TestOptimizer:
     r"""
-    Tests for the ``OptimizerJax`` class.
+    Tests for the ``Optimizer`` class.
     """
 
     @given(n=st.integers(0, 3))
@@ -61,7 +61,7 @@ class TestOptimizerJax:
         def cost_fn(S):
             return -(math.abs(S.fock_array((n + 1, n + 1))[n, n]) ** 2)
 
-        opt = OptimizerJax(euclidean_lr=0.01)
+        opt = Optimizer(euclidean_lr=0.01)
         (S,) = opt.minimize(cost_fn, by_optimizing=[S], max_steps=300)
 
         expected = 1 / (n + 1) * (n / (n + 1)) ** n
@@ -90,7 +90,7 @@ class TestOptimizerJax:
         def cost_fn(circ):
             return math.abs(circ.contract().fock_array((cutoff,) * 4)[i, 1, i + k - 1, k]) ** 2
 
-        opt = OptimizerJax(euclidean_lr=0.01)
+        opt = Optimizer(euclidean_lr=0.01)
         (circ,) = opt.minimize(
             cost_fn,
             by_optimizing=[circ],
@@ -122,7 +122,7 @@ class TestOptimizerJax:
             amps = circ.contract().fock_array((2, 2))
             return -(math.abs(amps[1, 1]) ** 2) + math.abs(amps[0, 1]) ** 2
 
-        opt = OptimizerJax(euclidean_lr=0.05)
+        opt = Optimizer(euclidean_lr=0.05)
 
         (circ,) = opt.minimize(cost_fn, by_optimizing=[circ], max_steps=300)
         assert math.allclose(-cost_fn(circ), 0.25, atol=1e-5)
@@ -135,7 +135,7 @@ class TestOptimizerJax:
             amps = G.fock_array((2, 2))
             return -(math.abs(amps[1, 1]) ** 2) + math.abs(amps[0, 1]) ** 2
 
-        opt = OptimizerJax(symplectic_lr=0.5, euclidean_lr=0.01)
+        opt = Optimizer(symplectic_lr=0.5, euclidean_lr=0.01)
 
         (G,) = opt.minimize(cost_fn, by_optimizing=[G], max_steps=500)
         assert math.allclose(-cost_fn(G), 0.25, atol=1e-4)
@@ -157,7 +157,7 @@ class TestOptimizerJax:
             amps = circ.contract().fock_array((2, 2))
             return -(math.abs(amps[1, 1]) ** 2) + math.abs(amps[0, 1]) ** 2
 
-        opt = OptimizerJax(unitary_lr=0.5, euclidean_lr=0.01)
+        opt = Optimizer(unitary_lr=0.5, euclidean_lr=0.01)
 
         (circ,) = opt.minimize(cost_fn, by_optimizing=[circ], max_steps=1000)
         assert math.allclose(-cost_fn(circ), 0.25, atol=1e-5)
@@ -187,7 +187,7 @@ class TestOptimizerJax:
             amps = circ.contract().fock_array((2, 2))
             return -(math.abs(amps[1, 1]) ** 2) + math.abs(amps[0, 1]) ** 2
 
-        opt = OptimizerJax(orthogonal_lr=0.5, euclidean_lr=0.01)
+        opt = Optimizer(orthogonal_lr=0.5, euclidean_lr=0.01)
 
         (circ,) = opt.minimize(cost_fn, by_optimizing=[circ], max_steps=1000)
         assert math.allclose(-cost_fn(circ), 0.25, atol=1e-5)
@@ -252,7 +252,7 @@ class TestOptimizerJax:
             amps = circ.contract().fock_array((3, 3, 3, 3))
             return -(math.abs((amps[1, 1, 2, 0] + amps[1, 1, 0, 2]) / np.sqrt(2)) ** 2)
 
-        opt = OptimizerJax(unitary_lr=0.05)
+        opt = Optimizer(unitary_lr=0.05)
         (circ,) = opt.minimize(cost_fn, by_optimizing=[circ], max_steps=200)
         assert math.allclose(-cost_fn(circ), 0.0625, atol=1e-5)
 
@@ -316,7 +316,7 @@ class TestOptimizerJax:
             amps = circ.contract().fock_array((2, 2, 3, 3))
             return -(math.abs((amps[1, 1, 0, 2] + amps[1, 1, 2, 0]) / np.sqrt(2)) ** 2)
 
-        opt = OptimizerJax()
+        opt = Optimizer()
 
         (circ,) = opt.minimize(cost_fn, by_optimizing=[circ], max_steps=200)
         assert math.allclose(-cost_fn(circ), 0.0625, atol=1e-5)
@@ -343,7 +343,7 @@ class TestOptimizerJax:
         def cost_fn(circ):
             return math.abs(circ.contract().fock_array((2, 2, 2, 2))[1, 1, 1, 1]) ** 2
 
-        opt = OptimizerJax(euclidean_lr=0.001)
+        opt = Optimizer(euclidean_lr=0.001)
         (circ,) = opt.minimize(cost_fn, by_optimizing=[circ], max_steps=300)
         S_12 = circ.components[3]
         assert math.allclose(math.sinh(S_12.parameters.r.value) ** 2, 1, atol=1e-2)
@@ -364,7 +364,7 @@ class TestOptimizerJax:
         def cost_fn(circ):
             return math.abs(circ.contract().fock_array((2, 2, 2, 2))[1, 1, 1, 1]) ** 2
 
-        opt = OptimizerJax(euclidean_lr=0.001)
+        opt = Optimizer(euclidean_lr=0.001)
         (circ,) = opt.minimize(cost_fn, by_optimizing=[circ], max_steps=300)
         r_var = circ.components[3].parameters.r
         assert math.allclose(math.sinh(r_var.value) ** 2, 1, atol=1e-2)
@@ -397,7 +397,7 @@ class TestOptimizerJax:
             entropy = von_neumann_entropy(cov0)
             return math.abs((num_mean0 - nbar) ** 2 + (entropy - S) ** 2 + (num_mean1 - nbar) ** 2)
 
-        opt = OptimizerJax(symplectic_lr=0.1)
+        opt = Optimizer(symplectic_lr=0.1)
         (G,) = opt.minimize(cost_fn, by_optimizing=[G], max_steps=50)
         S = math.asnumpy(G.parameters.symplectic.value)
         cov = S @ S.T
@@ -412,7 +412,7 @@ class TestOptimizerJax:
             state_out = Vacuum(0) >> dgate
             return -(math.abs(math.sum(math.conj(state_out.fock_array((40,))) * target_state)) ** 2)
 
-        opt = OptimizerJax()
+        opt = Optimizer()
         (dgate,) = opt.minimize(cost_fn, by_optimizing=[dgate])
 
         assert math.allclose(dgate.parameters.x.value, 0.1, atol=0.01)
@@ -427,7 +427,7 @@ class TestOptimizerJax:
             state_out = Vacuum(0) >> sgate
             return -(math.abs(math.sum(math.conj(state_out.fock_array((40,))) * target_state)) ** 2)
 
-        opt = OptimizerJax()
+        opt = Optimizer()
         (sgate,) = opt.minimize(cost_fn, by_optimizing=[sgate])
 
         assert math.allclose(sgate.parameters.r.value, 0.1, atol=0.01)
@@ -441,7 +441,7 @@ class TestOptimizerJax:
         def cost_fn(bsgate):
             return -(math.abs(math.sum(math.conj(bsgate.fock_array(40)) * target_gate)) ** 2)
 
-        opt = OptimizerJax()
+        opt = Optimizer()
         (bsgate,) = opt.minimize(cost_fn, by_optimizing=[bsgate])
 
         assert math.allclose(bsgate.parameters.theta.value, 0.1, atol=0.01)
@@ -459,7 +459,7 @@ class TestOptimizerJax:
             norm = 1 / squeezing.ansatz.batch_size if squeezing.ansatz.batch_shape else 1
             return -math.real(norm * math.sum(num >> squeezing >> vac) ** 2)
 
-        opt = OptimizerJax(euclidean_lr=0.05)
+        opt = Optimizer(euclidean_lr=0.05)
         (squeezing,) = opt.minimize(cost_fn, by_optimizing=[squeezing], max_steps=100)
 
         assert math.all(squeezing.parameters.r.value != og_r)
@@ -477,7 +477,7 @@ class TestOptimizerJax:
             norm = 1 / disp.ansatz.batch_size if disp.ansatz.batch_shape else 1
             return -math.real(norm * math.sum(num >> disp >> vac) ** 2)
 
-        opt = OptimizerJax(euclidean_lr=0.05)
+        opt = Optimizer(euclidean_lr=0.05)
         (disp,) = opt.minimize(cost_fn, by_optimizing=[disp], max_steps=100)
         assert math.all(og_x != disp.parameters.x.value)
         assert math.all(og_y != disp.parameters.y.value)
@@ -497,7 +497,7 @@ class TestOptimizerJax:
                 norm * math.sum(sq >> num >> bs >> (vac >> num).dual) ** 2,
             )
 
-        opt = OptimizerJax(euclidean_lr=0.05)
+        opt = Optimizer(euclidean_lr=0.05)
         (sq,) = opt.minimize(cost_fn, by_optimizing=[sq], max_steps=100)
 
         assert math.all(og_r != sq.parameters.r.value)
@@ -514,7 +514,7 @@ class TestOptimizerJax:
 
         # stable_threshold and max_steps are set to whatever gives us optimized parameters
         # that are within the default ATOL=1e-8 of the expected values
-        opt = OptimizerJax(stable_threshold=1e-12)
+        opt = Optimizer(stable_threshold=1e-12)
         (cat_state,) = opt.minimize(cost_fn, by_optimizing=[cat_state], max_steps=6000)
 
         assert math.allclose(cat_state.parameters.x.value, expected_cat.parameters.x.value)
